# docs/swagger.yaml
definitions:
...
  requestmodel.SellerEditProfile:
    properties:
      confirmpassword:
        type: string
      description:
        type: string
      email:
        type: string
      gstno:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        minLength: 4
        type: string
    required:
    - confirmpassword
    - description
    - email
    - gstno
    - id
    - name
    - password
    type: object
  requestmodel.SellerLogin:
    properties:
      email:
        type: string
      password:
        minLength: 4
        type: string
    type: object
  requestmodel.SellerSignup:
    properties:
      confirmpassword:
        type: string
      description:
        type: string
      email:
        type: string
      gstno:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        minLength: 4
        type: string
    required:
    - description
    - gstno
    - name
    type: object
...
  response.Response:
    properties:
      error: {}
      message:
        type: string
      result: {}
      status_code:
        type: integer
    type: object
...
info:
  contact:
    email: shahabazsulthan4@gmail.com
    name: 'Creator: Shahabaz Sulthan'
  description: "Laptop Lounge - \U0001F680 Your One-Stop Destination for Ultimate
    Laptop Shopping!\n\U0001F6D2\U0001F4BB Browse, compare, and buy top-notch laptops
    effortlessly.\nPowered by cutting-edge technology, we bring you a seamless shopping
    experience.\n\U0001F31F Dive into the future of laptop shopping with Laptop Lounge!
    \U0001F680\U0001F525"
  title: Laptop Lounge API
paths:
...
  /admin/sellers/block/{sellerID}:
    patch:
      consumes:
      - application/json
      description: Using this handler, admin can block a seller.
      parameters:
      - description: Seller ID in the URL path
        in: path
        name: sellerID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully blocked seller
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Block Seller
      tags:
      - Admin Seller Control
  /admin/sellers/getsellers:
    get:
      consumes:
      - application/json
      description: Using this handler, admin can get a list of sellers.
      parameters:
      - default: 1
        description: Page number for pagination (default 1)
        in: query
        name: page
        required: true
        type: integer
      - default: 5
        description: Number of items to return per page (default 5)
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of sellers
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Get Sellers
      tags:
      - Admin Seller Control
  /admin/sellers/pending:
    get:
      consumes:
      - application/json
      description: Using this handler, admin can get a list of pending sellers.
      parameters:
      - default: 1
        description: Page number for pagination (default 1)
        in: query
        name: page
        required: true
        type: integer
      - default: 5
        description: Number of items to return per page (default 5)
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of pending sellers
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Get Pending Sellers
      tags:
      - Admin Seller Control
  /admin/sellers/singleview/{sellerID}:
    get:
      consumes:
      - application/json
      description: Using this handler, admin can get details of a single seller.
      parameters:
      - description: Seller ID in the URL path
        in: path
        name: sellerID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved seller details
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Get Single Seller Details
      tags:
      - Admin Seller Control
  /admin/sellers/unblock/:sellerID:
    patch:
      consumes:
      - application/json
      description: Using this handler, admin can block a seller
      parameters:
      - description: Seller ID in the URL path
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Block Seller
      tags:
      - Admin Seller Control
  /admin/sellers/verify/{sellerID}:
    patch:
      consumes:
      - application/json
      description: Using this handler, admin can verify a seller.
      parameters:
      - description: Seller ID in the URL path
        in: path
        name: sellerID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully verified the seller
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Verify Seller
      tags:
      - Admin Seller Control
...
  /seller/{SellerID}:
    get:
      consumes:
      - application/json
      description: Retrieve details for the seller sales.
      parameters:
      - description: Seller ID in the URL path
        in: path
        name: SellerID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Details retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized. Authentication required.
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      - Refreshtoken: []
      summary: Get Seller Dashboard
      tags:
      - Seller
...
  /seller/login:
    post:
      consumes:
      - application/json
      description: Using this handler, a seller can log in.
      parameters:
      - description: Seller login details
        in: body
        name: Seller
        required: true
        schema:
          $ref: '#/definitions/requestmodel.SellerLogin'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged in
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Seller Login
      tags:
      - Seller
...
  /seller/profile/{SellerID}:
    get:
      consumes:
      - application/json
      description: Retrieve details of the seller's profile.
      parameters:
      - description: Seller ID in the URL path
        in: path
        name: SellerID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved the seller's profile
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Seller Profile
      tags:
      - Seller Profile
    patch:
      consumes:
      - application/json
      description: Update the seller's profile.
      parameters:
      - description: Seller ID in the URL path
        in: path
        name: SellerID
        required: true
        type: string
      - description: Seller profile details for updating
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/requestmodel.SellerEditProfile'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated the seller's profile
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update Seller Profile
      tags:
      - Seller Profile
...
  /seller/signup:
    post:
      consumes:
      - application/json
      description: Using this handler, a seller can sign up.
      parameters:
      - description: Seller signup details
        in: body
        name: Seller
        required: true
        schema:
          $ref: '#/definitions/requestmodel.SellerSignup'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully signed up
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Seller Signup
      tags:
      - Seller
...
securityDefinitions:
  BearerTokenAuth:
    description: 'Use your Bearer token for authentication. Example: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
  RefreshTokenAuth:
    description: 'Use your Refresh token to obtain a new Bearer token. Example: "{token}"'
    in: header
    name: Refreshtoken
    type: apiKey
swagger: "2.0"